+++ examples/detect-socket-failure/Gopkg.toml
@@ -1,0 +1,11 @@
+[[constraint]]
+  name = "github.com/bdlm/errors"
+  version = "0.1.1"
+
+[[constraint]]
+  name = "github.com/mkenney/go-chrome"
+  version = "1.0.0-rc6"
+
+[prune]
+  go-tests = true
+  unused-packages = true
