+++ examples/detect-socket-failure/README.md
