+++ examples/detect-socket-failure/main.go
@@ -1,4 +1,5 @@
 // +build ignore
+
 package main
 
 import (
