+++ examples/get-page-html/Gopkg.toml
@@ -1,0 +1,7 @@
+[[constraint]]
+  name = "github.com/mkenney/go-chrome"
+  version = "1.0.0-rc6"
+
+[prune]
+  go-tests = true
+  unused-packages = true
