+++ examples/issue-94/Gopkg.toml
@@ -1,0 +1,15 @@
+[[constraint]]
+  name = "github.com/bdlm/log"
+  version = "0.1.10"
+
+[[constraint]]
+  name = "github.com/mkenney/go-chrome"
+  version = "1.0.0-rc6"
+
+[[constraint]]
+  name = "github.com/sanity-io/litter"
+  version = "1.1.0"
+
+[prune]
+  go-tests = true
+  unused-packages = true
