+++ examples/screenshot-html-string/Gopkg.toml
@@ -1,0 +1,11 @@
+[[constraint]]
+  name = "github.com/bdlm/log"
+  version = "0.1.10"
+
+[[constraint]]
+  name = "github.com/mkenney/go-chrome"
+  version = "1.0.0-rc6"
+
+[prune]
+  go-tests = true
+  unused-packages = true
